model_api/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
whylogs==0.6.2
python-dotenv==0.19.0
pandas==1.3.2
joblib==1.0.1
flask_cors==3.0.10
flasgger==0.9.5
gunicorn==20.0.0
pydantic==1.8.2
Flask-Pydantic==0.8.0
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
flask-track-usage==2.0.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.6.0
slackclient==1.0.0
